implemented deletion in UI
TLDR, it fixes #4.
- UI Delete button is a toggle now
- UI bleeds red in delete mode, normal in normal mode
- click on slot deletes it, both from indexing and disk.

TODO: well, without full refactor we need to pass ParentPicker pointer as a field to each of the cards, going through Grid. It can be good, as almost every UI element has an access to a control class, BUT it is like "global" reference. Need to think how to store this in a more clever way.